Major Update to Card Information
This project had fallen far behind the main in terms of card information.  Bringing it up to date.  Many data errors should be fixed, and the latest set (Wave 5) is added.

(Not copying over placeholder cards used to allow users to add custom cards in this update.  May do so in a future one.)