Starting Card-Based Ability System
Adding a methods.py file to hold a dictionary with cards and corresponding functions.  This should let the AI system "draw" and "play" cards.  Also moving some of the underlying function calls to it to tidy up the automation.py file a bit.

Also adding a pierce variable so pierce can be supported beyond just Black pips.  This, along with bold and tough (whose capitalization is fixed in this commit) are currently being stored as Global variables.  This is far from ideal.  Will want to implement a proper solution in the future.