Introducing the activate() Function
Introducing a first draft of the activate() function which will match a provided card to an entry in the cardList variable (if possible) and execute any functions associated with that card.  Could potentially be expanded in a number of ways, including allowing fine-control over which functions for a card get executed.