Version Update
Incrasing version from 0.0.0.1 to 0.0.0.2 as all the major code changes needed for the next release have been implemented.  All that's needed before a "live" test release now is to assign AI cards/decks/actions and make a couple tweaks to call them.

Potential feature to add before a "live" release is the ability to ask users to pick a deck to play against.  This hasn't been added yet as so far there we don't have even one "real" deck.